working anthropic models + updated token counter
