+++ .env.example
@@ -5,4 +5,4 @@ REDDIT_CLIENT_SECRET=YOUR_CLIENT_SECRET
 REDDIT_USERNAME=YOUR_USERNAME
 REDDIT_PASSWORD=YOUR_PASSWORD
 REDDIT_USER_AGENT='linux:com.youragent.reddit-gpt-summarizer:v1.0.0 (by /u/yourusername)'
-ANTHROPIC_KEY=YOUR_ANTHROPIC_KEY
+ANTHROPIC_API_KEY=YOUR_ANTHROPIC_KEY
