+++ README.md
@@ -1,5 +1,13 @@
 # Reddit GPT Summarizer
 
+## 5/22/2023: added support for Anthropic models including Claude 100k + older OpenAI Instruct Models
+
+Expand settings to use Anthropic models; also added support for older OpenAI instruct models-- most produce garbage outputs but useful to test, that being said, Text Davinci 003 subjectively produces some of the highest quality outputs.  The new 100k models can often consume entire reddit threads without recursion.
+
+Don't forget to add you Anthropic API key to your .env file. (ANTHROPIC_API_KEY)
+
+https://www.anthropic.com/index/100k-context-windows
+
 ## 3/17/2023: added support for GPT-4 models
 
 If you have access to the API, you can use the longer context windows today. See docs.
