+++ app/env.py
@@ -23,6 +23,7 @@ class EnvVars(TypedDict):
     REDDIT_USERNAME: Optional[str]
     REDDIT_PASSWORD: Optional[str]
     REDDIT_USER_AGENT: str
+    ANTHROPIC_API_KEY: str
 
 
 class EnvVarsLoader:
@@ -56,6 +57,7 @@ class EnvVarsLoader:
         reddit_username = os.getenv("REDDIT_USERNAME")
         reddit_password = os.getenv("REDDIT_PASSWORD")
         reddit_user_agent = os.getenv("REDDIT_USER_AGENT")
+        anthropic_api_key = os.environ["ANTHROPIC_API_KEY"]
 
         if org_id is None or api_key is None:
             err_msg = "Missing OpenAI API key or organization ID."
@@ -79,6 +81,7 @@ class EnvVarsLoader:
             "REDDIT_USERNAME": reddit_username,
             "REDDIT_PASSWORD": reddit_password,
             "REDDIT_USER_AGENT": reddit_user_agent,
+            "ANTHROPIC_API_KEY": anthropic_api_key,
         }
 
         EnvVarsLoader._env_vars = env_vars
