+++ app/tests/test_utils.py
@@ -2,8 +2,9 @@
 import os
 
 import tiktoken
-from services.openai_methods import num_tokens_from_string
+from config import OPEN_AI_CHAT_TYPE
 from utils.common import generate_filename, get_timestamp, save_output
+from utils.llm_utils import num_tokens_from_string
 
 
 def test_num_tokens_from_string() -> None:
@@ -12,7 +13,7 @@ def test_num_tokens_from_string() -> None:
     encoding_name: str = "gpt2"
 
     tiktoken.get_encoding(encoding_name)
-    result: int = num_tokens_from_string(string, encoding_name)
+    result: int = num_tokens_from_string(string, OPEN_AI_CHAT_TYPE)
 
     assert result == 3
