+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "reddit-gpt-summarizer"
-version = "0.1.0"
-description = "Summarize Reddit Threads with GPT"
+version = "0.2.0"
+description = "Summarize Reddit Threads with LLMs"
 authors = ["Sean Dearnaley <seandearnaley@hotmail.com>"]
 readme = "README.md"
 
@@ -15,6 +15,7 @@ streamlit = "^1.18.1"
 pyrate-limiter = "^2.10.0"
 praw = "^7.7.0"
 colorlog = "^6.7.0"
+anthropic = "^0.2.9"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
